refactor: move Table testing to test_table.py
